Add initial configuration file package
This commit adds the configuration of conf2el in a separate Go package called
"config". It represents all fields in structs, and every struct comes with a
Validate() function that will attempt to find errors in the data. All
Validate() functions come with unit tests that do not yet have 100% coverage.
The validators also don't check for every possible error yet.

The struct fields are not complete yet. For example, the Config struct lacks
the eldim server configuration that's required to upload data to.